requirements.txt
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
